Abort if deallocation fails in random number engines
Abort instead of silently continuing when `hipFree` fails in rng
destructors. `hipFree` should only fail if
1. there is something wrong
with the runtime in which case continuing doesn't really make sense or
2. There is a bug in rocRAND trying to double free or free something
that was not allocated with `hipMalloc`.

This is usually a sign that something went pretty wrong, and probably
other weird failures would follow if we tried to continue anyway.

This makes a test for move construction / assignment in the next commit
sound, so that it does not need to depend on implicitly terminating when
device side memory is used after freeing it.